refactor: code improvement, according to Sonar Quality requirements;
