+++ jans-auth-server/model/src/main/java/io/jans/as/model/crypto/signature/ECDSAKeyFactory.java
@@ -12,31 +12,18 @@ import io.jans.util.security.SecurityProviderUtility;
 
 import org.apache.commons.lang.StringUtils;
 
-import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.cert.X509CertificateHolder;
-import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter;
-import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
 import org.bouncycastle.operator.OperatorCreationException;
-import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
-
-import org.bouncycastle.asn1.ASN1EncodableVector;
-import org.bouncycastle.asn1.ASN1ObjectIdentifier;
-import org.bouncycastle.asn1.DERSequence;
-import org.bouncycastle.asn1.x509.KeyPurposeId;
 import org.bouncycastle.cert.CertIOException;
 
 import java.math.BigInteger;
 import java.security.AlgorithmParameters;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidParameterException;
-import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
 import java.util.Calendar;
-import java.util.Date;
 import java.util.GregorianCalendar;
 
 import java.security.interfaces.ECPrivateKey;
@@ -56,12 +43,8 @@ import java.security.spec.InvalidParameterSpecException;
  */
 public class ECDSAKeyFactory extends KeyFactory<ECDSAPrivateKey, ECDSAPublicKey> {
 
-    private final SignatureAlgorithm signatureAlgorithm;
-    private final KeyPair keyPair;
-
     private final ECDSAPrivateKey ecdsaPrivateKey;
     private final ECDSAPublicKey ecdsaPublicKey;
-    private Certificate certificate;
 
     public ECDSAKeyFactory(SignatureAlgorithm signatureAlgorithm, String dnName) throws NoSuchAlgorithmException, InvalidParameterSpecException, InvalidAlgorithmParameterException, OperatorCreationException, CertificateException, CertIOException {
         if (signatureAlgorithm == null) {
@@ -98,26 +81,6 @@ public class ECDSAKeyFactory extends KeyFactory<ECDSAPrivateKey, ECDSAPublicKey>
 
             this.certificate = generateV3Certificate(startDate.getTime(), expiryDate.getTime(), dnName);
         }
-    }
-
-    public Certificate generateV3Certificate(Date startDate, Date expirationDate, String dnName) throws OperatorCreationException, CertificateException, CertIOException {
-        BigInteger serialNumber = new BigInteger(1024, new SecureRandom()); // serial number for certificate
-        X500Name name = new X500Name(dnName);
-
-        JcaX509v3CertificateBuilder certGen = new JcaX509v3CertificateBuilder(name, serialNumber, startDate, expirationDate, name, keyPair.getPublic());
-
-        ASN1EncodableVector purposes = new ASN1EncodableVector();
-        purposes.add(KeyPurposeId.id_kp_serverAuth);
-        purposes.add(KeyPurposeId.id_kp_clientAuth);
-        purposes.add(KeyPurposeId.anyExtendedKeyUsage);
-
-        ASN1ObjectIdentifier extendedKeyUsage = new ASN1ObjectIdentifier("2.5.29.37").intern();
-        certGen.addExtension(extendedKeyUsage, false, new DERSequence(purposes));
-
-        X509CertificateHolder certHolder = certGen.build(new JcaContentSignerBuilder(signatureAlgorithm.getAlgorithm()).setProvider(SecurityProviderUtility.getBCProviderName()).build(keyPair.getPrivate()));
-        X509Certificate x509Certificate = new JcaX509CertificateConverter().setProvider(SecurityProviderUtility.getBCProviderName()).getCertificate(certHolder);
-
-        return new Certificate(signatureAlgorithm, x509Certificate);
     }
 
     @Override
